// src/app/models/starship.model.ts
export interface StarshipModel{
    model: string;
    crew: number | null;
    starshipClass: string;
    manufacturer: string;
    costInCredits: string;
    length: string;
    passengers: string;
    maxAtmospheringSpeed: string;
    hyperdriveRating: string;
    mglt: string;
    cargoCapacity: string;
    created: string
    edited: string;
    description: string;
}